requirements.txt
brotli==1.1.0
ConfigParser==6.0.0
cryptography==41.0.4
Cython==3.0.2
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
importlib_metadata==6.8.0
keyring==24.2.0
langchain==0.0.308
lockfile==0.12.2
lxml==4.9.3
numpy==1.24.3
ordereddict==1.1
pandas==2.1.1
protobuf==4.24.4
pyOpenSSL==23.2.0
python-dotenv==1.0.0
streamlit==1.27.1
streamlit_chat==0.1.1
toml==0.10.2
tornado==6.3.3
xmlrpclib==1.0.1
zipp==3.16.2
